Send emails from the contact form
Adds the Flask-Mail package from mattupstate, http://pythonhosted.org/Flask-Mail/

Sends all emails entered into the contact form to those admins who have been specified as mail recipients in the admin section.

Does not currently do any kind of anti-spam. The destination addresses are few and not publicly-configurable, so it's initially expected that our spam filters can take it. WTForms has built-in support for reCaptcha, so that can be added when it becomes an issue.